requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
Flask-Cors==3.0.10
python-socketio==5.3.0
python-engineio==4.3.4
mss==9.0.1
Pillow==10.1.0
werkzeug>=2.0.0,<3.0.0
requests==2.31.0
pywin32==306
+ opencv-python-headless==4.8.1.78
+ numpy==1.24.3